requirements.txt
langchain
streamlit
openai
qdrant-client
grpcio-tools
python-dotenv
tiktoken
langchain-community